Начало разработки напоминаний. Перенос логики запроса погоды на backend.

+++ frontend/src/pages/info/components/notes/components/note-row/Note-row.jsx
@@ -3,7 +3,7 @@ import { useDispatch } from 'react-redux';
 import { Link } from 'react-router-dom';
 import { Icon } from '../../../../../../components';
 import { removeNote } from '../../../../../../redux/actions';
-import { formatDate } from '../../../../../../utils';
+import { confirmed, formatDate } from '../../../../../../utils';
 import { PROP_TYPE } from '../../../../../../constans';
 import styles from './Note-row.module.css';
 
@@ -11,8 +11,11 @@ export const NotesRow = ({ note, users, handleNoteDelete }) => {
 	const dispatch = useDispatch();
 
 	const onNoteRemove = () => {
-		dispatch(removeNote(note.id));
-		handleNoteDelete(note.id);
+		if (confirmed('note')) {
+			dispatch(removeNote(note.id));
+			handleNoteDelete(note.id);
+		}
+		return;
 	};
 
 	return (
